+++ src/commands/index.js
@@ -29,6 +29,9 @@ import deleterule from './deleterule.js';
 import exportrules from './exportrules.js';
 import kick from './kick.js';
 import tempban from './tempban.js';
+import reportbug from './reportbug.js';
+import bugs from './bugs.js';
+import bugstatus from './bugstatus.js';
 
 const modules = [
  warn,
@@ -62,6 +65,9 @@ const modules = [
  exportrules,
  kick,
  tempban,
+ reportbug,
+ bugs,
+ bugstatus,
 ];
 
 export const commandRegistry = new Map();